+++ experimental/packages/opentelemetry-instrumentation-http/src/http.ts
@@ -30,10 +30,14 @@ import {
   ValueType,
 } from '@opentelemetry/api';
 import {
+  getStringListFromEnv,
   hrTime,
   hrTimeDuration,
   hrTimeToMilliseconds,
   suppressTracing,
+  RPCMetadata,
+  RPCType,
+  setRPCMetadata,
 } from '@opentelemetry/core';
 import type * as http from 'http';
 import type * as https from 'https';
@@ -46,12 +50,6 @@ import {
   InstrumentationNodeModuleDefinition,
   safeExecuteInTheMiddle,
 } from '@opentelemetry/instrumentation';
-import {
-  RPCMetadata,
-  RPCType,
-  setRPCMetadata,
-  getEnv,
-} from '@opentelemetry/core';
 import { errorMonitor } from 'events';
 import {
   ATTR_HTTP_REQUEST_METHOD,
@@ -97,10 +95,10 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
   /** keep track on spans not ended */
   private readonly _spanNotEnded: WeakSet<Span> = new WeakSet<Span>();
   private _headerCapture;
-  private _oldHttpServerDurationHistogram!: Histogram;
-  private _stableHttpServerDurationHistogram!: Histogram;
-  private _oldHttpClientDurationHistogram!: Histogram;
-  private _stableHttpClientDurationHistogram!: Histogram;
+  declare private _oldHttpServerDurationHistogram: Histogram;
+  declare private _stableHttpServerDurationHistogram: Histogram;
+  declare private _oldHttpClientDurationHistogram: Histogram;
+  declare private _stableHttpClientDurationHistogram: Histogram;
 
   private _semconvStability = SemconvStability.OLD;
 
@@ -108,7 +106,8 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
     super('@opentelemetry/instrumentation-http', VERSION, config);
     this._headerCapture = this._createHeaderCapture();
 
-    for (const entry of getEnv().OTEL_SEMCONV_STABILITY_OPT_IN) {
+    for (const entry of getStringListFromEnv('OTEL_SEMCONV_STABILITY_OPT_IN') ??
+      []) {
       if (entry.toLowerCase() === 'http/dup') {
         // http/dup takes highest precedence. If it is found, there is no need to read the rest of the list
         this._semconvStability = SemconvStability.DUPLICATE;