+++ packages/opentelemetry-sdk-trace-base/test/common/Span.test.ts
@@ -27,8 +27,6 @@ import {
   AttributeValue,
 } from '@opentelemetry/api';
 import {
-  DEFAULT_ATTRIBUTE_COUNT_LIMIT,
-  DEFAULT_ATTRIBUTE_VALUE_LENGTH_LIMIT,
   hrTimeDuration,
   hrTimeToMilliseconds,
   hrTimeToNanoseconds,
@@ -45,6 +43,10 @@ import { BasicTracerProvider, Span, SpanProcessor } from '../../src';
 import { SpanImpl } from '../../src/Span';
 import { invalidAttributes, validAttributes } from './util';
 import { Tracer } from '../../src/Tracer';
+import {
+  DEFAULT_ATTRIBUTE_COUNT_LIMIT,
+  DEFAULT_ATTRIBUTE_VALUE_LENGTH_LIMIT,
+} from '../../src/utility';
 
 const performanceTimeOrigin: HrTime = [1, 1];
 
@@ -1005,14 +1007,18 @@ describe('Span', () => {
       spanContext,
       name: 'my-span',
       kind: SpanKind.INTERNAL,
-      parentSpanId: parentId,
+      parentSpanContext: {
+        spanId: parentId,
+        traceId: '',
+        traceFlags: TraceFlags.SAMPLED,
+      },
       spanLimits: tracer.getSpanLimits(),
       spanProcessor: tracer['_spanProcessor'],
     });
 
     assert.strictEqual(span.name, 'my-span');
     assert.strictEqual(span.kind, SpanKind.INTERNAL);
-    assert.strictEqual(span.parentSpanId, parentId);
+    assert.strictEqual(span.parentSpanContext?.spanId, parentId);
     assert.strictEqual(span.spanContext().traceId, spanContext.traceId);
     assert.deepStrictEqual(span.status, {
       code: SpanStatusCode.UNSET,